Fix for speed

+++ final_task/rss_reader/converting.py
@@ -7,6 +7,9 @@ import requests as r
 import printers
 import check
 import os
+import sys
+import re
+import random
 
 
 def create_html(items: list) -> document:
@@ -42,8 +45,6 @@ def create_pdf(items: list) -> None:
     pdf.write(8, "RSS feed")
     pdf.ln(20)
     for item in items:
-        b = 0
-        b = b + 1
         item = printers.prepare_one_item(item)
         pdf.write(8, "===Wow! News!===")
         pdf.ln(10)
@@ -51,15 +52,18 @@ def create_pdf(items: list) -> None:
         pdf.ln(10)
         pdf.write(8, "Link: " + str(item['Link: ']))
         pdf.ln(15)
-        if check.internet_on:
-            img_url = str(item['Media content:\n'])
-            if img_url != '':
-                image = r.get(img_url)
-                img_path = str(b) + '.' + str(img_url[-3:])
-                with open(img_path, 'wb') as file:
-                    file.write(image.content)
-                    pdf.image(img_path, w=70, h=50)
+        img_url = str(item['Media content:\n'])
+        url_list = img_url.split("\n")
+        if check.internet_on and img_url != '':
+            for element in url_list:
+                if element.endswith(".png") or element.endswith(".jpg"):
+                    image = r.get(element)
+                    img_path = str(len(element)) + str(random.randint(1, 128)) + str(img_url[-4:])
+                    with open(img_path, 'wb') as file:
+                        file.write(image.content)
+                        pdf.image(img_path, w=70, h=50)
                 os.remove(img_path)
+
         else:
             pdf.write(8, "Media content: " + str(item['Media content:\n']))
         pdf.ln(15)
